// src/app/interfaces/participante.interface.ts
export class Participante {
    id: number;
    nombre: string;
    apellido: string;
    puesto: string;
    oficinas: string;
    estado: string;
    correo: string;
    telefono: string;
    politica: number;
    email: string;
    empresa: string;
    monto: number;
    codigoDescuento: string;
    formaPago: string;
    confirmacionPaypal: string;
    referencia: string;
    rfc: string;
    razonSocial: string;
    calle: string;
    numeroExterior: string;
    numeroInterior: string;
    municipio: string;
    colonia: string;
    cp: string;
    correoFactura: string;
    usoCFDI: string;
    formaPagoFiscal: string;
    fechaRegistro: string;
    estatusRegistro: string;
    idGrupo: number;
    total: number;
    nombreInscribe: string;
    apellidoInscribe: string;
    correoInscribe: string;
    telefonoInscribe: string;
    emailFactura: string;
    idInscribe: number;
}